comments related to post

+++ src/modules/comments/services/comments.service.ts
@@ -1,15 +1,17 @@
 import { Injectable, NotFoundException } from "@nestjs/common";
 import { InjectModel } from "@nestjs/mongoose";
-import { Model } from "mongoose";
+import { Model, isValidObjectId } from 'mongoose';
 import { User } from "../../users/schemas/user.schema";
 import { CommentDto, EditCommentDto } from "../dtos/comment.dtos";
 import { Comment } from "../entities/comment.entity";
+import { Post } from '../../posts/schemas/post.schema';
 
 @Injectable()
 export class CommentsService {
   constructor(
     @InjectModel("Comments") private readonly commentModel: Model<Comment>,
-    @InjectModel("Users") private readonly userModel: Model<User>
+    @InjectModel("Users") private readonly userModel: Model<User>,
+    @InjectModel("Posts") private readonly postModel: Model<Post>,
   ) {}
 
   async findAll(): Promise<Comment[]> {
@@ -27,19 +29,47 @@ export class CommentsService {
     return comment;
   }
 
-  async create(data: CommentDto): Promise<Comment> {
-    const comment = new this.commentModel(data);
-    return await comment.save();
+  async create(data: CommentDto, ownerUserId:string): Promise<Comment> {
+    if(!isValidObjectId(ownerUserId) || !isValidObjectId(data.postId)) throw new NotFoundException(`Ids not valid`);
+
+    const findUser = await this.userModel.findById(ownerUserId);
+    const findPost = await this.postModel.findById(data.postId);
+    
+    if(!findUser) throw new NotFoundException(`The user with the ID: '${ownerUserId}' was not found.`);
+    if(!findPost) throw new NotFoundException(`The post with the ID: '${data.postId}' was not found.`);
+
+    const comment = await new this.commentModel({
+      comment: data.comment,
+      user: ownerUserId,
+    }).save()
+
+    await findPost.updateOne({
+      $push: {
+        comments: comment.id
+      }
+    });
+
+    return comment;
   }
 
-  async update(id: string, comment: EditCommentDto): Promise<Comment> {
-    const commentUpdate = await this.commentModel.findByIdAndUpdate(id, comment, {
-      new: true
-    });
+  async update(id: string, data: EditCommentDto): Promise<Comment> {
+    if(!isValidObjectId(id)) throw new NotFoundException(`Id not valid`);
+    
+    const findComment = await this.commentModel.findById(id);
+    if(!findComment) throw new NotFoundException(`The comment with the ID: '${id}' was not found.`);
+
+    const commentUpdate = await this.commentModel.findByIdAndUpdate(id, {
+      comment: data.comment
+    }, { new: true });
     return commentUpdate;
   }
 
   async delete(id: string): Promise<void> {
+    if(!isValidObjectId(id)) throw new NotFoundException(`Id not valid`);
+
+    const findComment = await this.commentModel.findById(id);
+    if(!findComment) throw new NotFoundException(`The comment with the ID: '${id}' was not found.`);
+    
     await this.commentModel.findByIdAndRemove(id);
   }
 }
